Expose constructor to subclasses
To give users of NetflixOSS a chance to modify the LoggingConfiguration
(e.g. not use log4j) we need to expose at least a constructor, so that
a subclass can be created that overrides some of the default behaviour.
More could be done to make the logging more flexible, but if we start
with this then one can make some progress.